Refactor - move Vs independent code to ProjectSystem.Managed project
